Renamed the package from com.arhimisha.intech.domain to com.arhimisha.intech.dto
