+++ src/main/java/com/arhimisha/intech/controllers/MainController.java
@@ -1,11 +1,10 @@
 package com.arhimisha.intech.controllers;
 
-import com.arhimisha.intech.domain.User;
+import com.arhimisha.intech.dto.User;
 import com.arhimisha.intech.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
