+++ src/main/java/com/arhimisha/intech/controllers/SubjectController.java
@@ -1,7 +1,7 @@
 package com.arhimisha.intech.controllers;
 
-import com.arhimisha.intech.domain.Subject;
-import com.arhimisha.intech.domain.User;
+import com.arhimisha.intech.dto.Subject;
+import com.arhimisha.intech.dto.User;
 import com.arhimisha.intech.services.MessageService;
 import com.arhimisha.intech.services.SubjectService;
 import com.arhimisha.intech.services.UserService;
