+++ src/main/java/com/arhimisha/intech/dto/Authority.java
@@ -1,4 +1,4 @@
-package com.arhimisha.intech.domain;
+package com.arhimisha.intech.dto;
 
 import org.springframework.security.core.GrantedAuthority;
