+++ src/main/java/com/arhimisha/intech/dto/User.java
@@ -1,4 +1,4 @@
-package com.arhimisha.intech.domain;
+package com.arhimisha.intech.dto;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
