+++ src/main/java/com/arhimisha/intech/repository/SubjectRepository.java
@@ -1,6 +1,6 @@
 package com.arhimisha.intech.repository;
 
-import com.arhimisha.intech.domain.Subject;
+import com.arhimisha.intech.dto.Subject;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
