+++ src/main/java/com/arhimisha/intech/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.arhimisha.intech.repository;
 
-import com.arhimisha.intech.domain.User;
+import com.arhimisha.intech.dto.User;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
